+++ src/grpc_server.cc
@@ -271,5 +271,43 @@ namespace grpc_labview
             _server = nullptr;
         }
         grpc_labview::ProtoDescriptorString::getInstance()->deleteInstance();
-    }    
+    }
+
+    //---------------------------------------------------------------------
+    //---------------------------------------------------------------------
+
+    // Initialize the static members
+    ProtoDescriptorString* ProtoDescriptorString::m_instance = nullptr;
+    std::mutex ProtoDescriptorString::m_mutex;
+
+    // Return the static class instance. Thread safe.
+    ProtoDescriptorString* ProtoDescriptorString::getInstance() {
+        std::unique_lock<std::mutex> lock(m_mutex);
+        if (m_instance == nullptr) {
+            m_instance = new ProtoDescriptorString();
+        }
+        return m_instance;
+    }
+
+    // Get the descriptor string
+    std::string ProtoDescriptorString::getDescriptor() {
+        std::unique_lock<std::mutex> lock(m_mutex);
+        return m_descriptor;
+    }
+
+    // Set the descriptor string
+    void ProtoDescriptorString::setDescriptor(std::string str) {
+        std::unique_lock<std::mutex> lock(m_mutex);
+        m_refcount++;
+        m_descriptor = str;
+    }
+
+    // Delete the instaance based on the refcount
+    void ProtoDescriptorString::deleteInstance() {
+        std::unique_lock<std::mutex> lock(m_mutex);
+        if (!--m_refcount) {
+            delete m_instance;
+            m_instance = nullptr;
+        }
+    }
 }
